Add CanvasPanel component
