+++ src/components/App.js
@@ -3,9 +3,7 @@ import * as React from 'react'
 import { connect } from 'react-redux'
 
 import style from './app.css'
-import Canvas from './Canvas'
-import Panel from './Panel'
-import SelectPalette from './SelectPalette'
+import CanvasPanel from './CanvasPanel'
 import TilePanel from './TilePanel'
 import Workspace from './Workspace'
 import { clearTile, resetWorkspace } from '../actions'
@@ -23,42 +21,21 @@ type DispatchProps = {
 
 type Props = DispatchProps & MappedProps
 
-const App = (props: Props) => {
-  const onClear = () => {
-    if (props.activeTileID) {
-      props.clearTile(props.activeTileID)
-    }
-  }
+const App = (props: Props) => (
+  <div className={style.app}>
+    <header className={style.header}>
+      <h1>GameBoy Z80 Tools</h1>
+      <button onClick={props.resetWorkspace} type="button">
+        Reset Layout
+      </button>
+    </header>
 
-  return (
-    <div className={style.app}>
-      <header className={style.header}>
-        <h1>GameBoy Z80 Tools</h1>
-        <button onClick={props.resetWorkspace} type="button">
-          Reset Layout
-        </button>
-      </header>
-
-      <Workspace>
-        <Panel id="CanvasPanel" title="Tile Canvas">
-          <div className={style.canvas}>
-            <div className={style.controls}>
-              <SelectPalette />
-              <div>
-                <button onClick={onClear} type="button">
-                  Clear
-                </button>
-              </div>
-            </div>
-            <Canvas height={8} width={8} />
-          </div>
-        </Panel>
-
-        <TilePanel />
-      </Workspace>
-    </div>
-  )
-}
+    <Workspace>
+      <CanvasPanel />
+      <TilePanel />
+    </Workspace>
+  </div>
+)
 
 const mapState = (state: AppState) => ({
   activeTileID: getActiveTileID(state),
