+++ src/components/app.css
@@ -52,17 +52,3 @@ h1 {
   margin: 0 auto;
   max-width: 960px;
 }
-
-.canvas {
-  align-items: center;
-  display: flex;
-  flex: 1 1 100%;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.controls {
-  align-items: center;
-  display: flex;
-  justify-content: space-between;
-}
